pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "lattice-tracker"
version = "0.1.0"
description = "File-based, agent-native task tracker with an event-sourced core."
requires-python = ">=3.12"
dependencies = [
    "click>=8.1",
    "python-ulid>=2.0",
    "filelock>=3.13",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "ruff>=0.4",
]

[project.scripts]
lattice = "lattice.cli.main:cli"

[tool.hatch.build.targets.wheel]
packages = ["src/lattice"]

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.ruff]
src = ["src"]
line-length = 99
